+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added multiple tabs to admin page
+
 ## [1.12.0] - 2022-05-16
 
 ### Added